requirements.txt
flask>=3.0
pdfplumber>=0.10
gunicorn>=21.2
psycopg2-binary>=2.9
gunicorn>=21.0
